Series of checks to abort installation if any errors are encounted

+++ installer/src/Project/Theme/Downloader.php
@@ -2,8 +2,10 @@
 
 namespace Message\Mothership\Install\Project\Theme;
 
+use Message\Mothership\Install\Exception\InstallFailedException;
 use Message\Mothership\Install\Output\InfoOutput;
 use Message\Mothership\Install\FileSystem\DirectoryResolver;
+use Message\Mothership\Install\FileSystem\FileResolver;
 use Message\Mothership\Install\Command\ShellCommand;
 
 /**
@@ -22,14 +24,20 @@ class Downloader
 	private $_dirResolver;
 
 	/**
+	 * @var \Message\Mothership\Install\FileSystem\FileResolver
+	 */
+	private $_fileResolver;
+
+	/**
 	 * @var \Message\Mothership\Install\Output\InfoOutput
 	 */
 	private $_info;
 
 	public function __construct()
 	{
-		$this->_dirResolver = new DirectoryResolver;
-		$this->_info        = new InfoOutput;
+		$this->_dirResolver  = new DirectoryResolver;
+		$this->_fileResolver = new FileResolver;
+		$this->_info         = new InfoOutput;
 	}
 
 	/**
@@ -40,6 +48,10 @@ class Downloader
 	 */
 	public function download(ThemeInterface $theme, $path)
 	{
+		if (!$this->_dirResolver->isEmpty($path)) {
+			throw new InstallFailedException('Could not download theme, the target directory `' . $path . '` must be empty');
+		}
+
 		$path = $this->_dirResolver->getAbsolute($path);
 
 		if (!$this->_dirResolver->exists($path)) {
@@ -49,7 +61,17 @@ class Downloader
 
 		$this->_info->info('Downloading theme, this may take a while');
 		ShellCommand::run('git clone -b ' . $theme->getBranch() . ' --single-branch '  . $theme->getGitRepo() . ' ' . $path);
+
+		if (!$this->_themeDownloaded($path)) {
+			throw new InstallFailedException('Theme not downloaded to `' . $path . '`, aborting install');
+		}
+
 		$this->_dirResolver->delete($path . '/.git');
 		$this->_dirResolver->delete($path . '/.gitignore');
+	}
+
+	private function _themeDownloaded($path)
+	{
+		return $this->_fileResolver->exists(rtrim($path, '/') . '/.git');
 	}
 }
